16 bit modifications

See merge request parallel-programming/MPR!14

+++ tests/cpp-tests/data-units/TestMPRTile.cpp
@@ -157,7 +157,7 @@ TEST_MPR_TILE() {
 
         precisions.resize(precision_size);
         for (auto i = 0; i < precision_size; i++) {
-            precisions[ i ] = "int";
+            precisions[ i ] = "half";
         }
         MPRTile a(10, 12, 2, 4, data, precisions);
